gitignore, scripts: Update promo pricing and products_old csv
Update the promotion pricing for shoprite and checkers to ensure
the format matches the normal price with the "R" prefixed.
Update the products_old.csv such that invalid shoprite image_urls
are removed which will result in them being scraped again.
Add the checkers and shoprite images folders to the gitignore file.

Signed-off-by: Joshua Van Straaten <joshuavanstraaten100@gmail.com>